paper doll wip
let's write a companion monobehaviour to this Dancer monobehaviour, called PaperDoll. currently, there is a 3d pose getting drawn for this dancer for every frame of the animation.

let's also load fig1_poses2d.json and fig2_poses2d.json for each dancer (Lead is fig1 and Follow is fig2) from the HeadMovement script and pass it to the respective dancer. the pose data looks like this:

[
    [
        [
            57,
            392
        ],
        [
            83,
            427
        ],
        [
            34,
            430
        ],
        [
            50,
            350
        ],
        [
            97,

etc

it is a list of list of integer pairs that describe the SMPL 2D pose locations (24 points) that correspond to the png of the dancer at that frame

the pngs of the dancer are at a folder of images, where each frame is formatted ####.png

when the dancer is created, the PaperDoll monobehaviour needs to be a RequiredComponent. In the awake of the PaperDoll, we need to create a quad that can receive a transparent png which will be loaded and assigned at every frame

when the dancer is set to frame N, we need to load the image for Lead (fig1) or Follow (fig2) from the folder, get the 2d pose for that frame, and then assign the image to the dancer image texture.

finally, the left shoulder of the person in the image needs to be pinned to the left shoulder of the 3d pose. this is accomplished by taking the upper left 2d shoulder pose (an integer xy position on the png) and setting the location of the image to correspond with the 3d left shoulder position. a ray needs to be drawn from world position 0,1.2,0 (the camera height) to the right hip of the 3d pose. the image needs to be uniformly scaled and rotated so that the 2d xy position of the right hip of the dancer in the image overlaps with this 3d position

please fully implement this PaperDoll monobehaviour, and load the necessary data from HeadMovement into Dancer, and make PaperDoll a required component of Dancer